Refactored project structure for consistency.
