--- Dataset/Coronaviruses/Host proteins (potential drug targets).csv
@@ -1,129 +1,0 @@
-Coronavirus ,Alternative virus name,Host Protein ,UniProt ID,Host Gene Entrez ID ,PubMed ID/STRING
-IBV ,IBV ,NONO ,Q15233,4841,23637410
-IBV ,IBV ,GSK3A,P49840,2931,23637410
-IBV ,IBV ,GSK3B,P49841,2932,23637410
-IBV ,IBV ,PABPC1,P11940,26986,23637410
-IBV ,IBV ,PABPC4,Q13310,8761,23637410
-IBV ,IBV ,HNRNPA1,P09651,3178,23637410
-IBV ,IBV ,HNRNPA2B1,P22626,3181,23637410
-IBV ,IBV ,NPM1,P06748,4869,23637410
-IBV ,IBV ,G3BP1,Q13283,10146,23637410
-IBV ,IBV ,G3BP2,Q9UN86,9908,23637410
-IBV ,IBV ,RPL19,P84098,6143,23637410
-IBV ,IBV ,PARP1,P09874,142,23637410
-IBV ,IBV ,NCL ,P19338,4691,23637410
-IBV ,IBV ,DDX1,Q92499,1653,23637410
-SARS ,SARS-CoV[GZ02],RYBP ,Q8N488,23429,22046132
-SARS ,SARS-CoV[GZ02],PPIA ,P62937,5478,22046132
-SARS ,SARS-CoV[GZ02],NOMO3,P69849,408050,22046132
-SARS ,SARS-CoV[GZ02],FKBP1A,P62942,2280,22046132
-SARS ,SARS-CoV[GZ02],PPIG ,Q13427,9360,22046132
-SARS ,SARS-CoV[GZ02],MARK3,P27448,4140,22046132
-SARS ,SARS-CoV[GZ02],PPIH ,O43447,10465,22046132
-SARS ,SARS-CoV[GZ02],RCAN3,Q9UKA8,11123,22046132
-SARS ,SARS-CoV[GZ02],HGS ,O14964,9146,22046132
-SARS ,SARS-CoV[GZ02],BAG6,P46379,7917,22046132
-SARS ,SARS-CoV[GZ02],DDAH2,O95865,23564,22046132
-SARS ,SARS-CoV[GZ02],CAMLG ,P49069,819,22046132
-SARS ,SARS-CoV[GZ02],CHMP2B,Q9UQN3,25978,22046132
-SARS ,SARS-CoV[GZ02],SNAP47,Q5SQN1,116841,22046132
-SARS ,SARS-CoV[GZ02],MKRN2,Q9H000,23609,22046132
-SARS ,SARS-CoV[GZ02],TPSAB1,Q15661,7177,22046132
-SARS ,SARS-CoV[GZ02],SERPING1,P05155,710,22046132
-SARS ,SARS-CoV[GZ02],MKRN3,Q13064,7681,22046132
-SARS ,SARS-CoV[GZ02],PSMA2,P25787,5683,22046132
-SARS ,SARS-CoV[GZ02],ABHD17A,Q96GS6,81926,22046132
-SARS ,SARS-CoV[GZ02],PFDN5,Q99471,5204,22046132
-SARS ,SARS-CoV[GZ02],MIF4GD,A9UHW6,57409,22046132
-SARS ,SARS-CoV[GZ02],NDUFA10,O95299,4705,22046132
-SARS ,SARS-CoV[GZ02],VKORC1,Q9BQB6,79001,22046132
-SARS ,SARS-CoV[GZ02],LAS1L,Q9Y4W2,81887,22046132
-SARS ,SARS-CoV[GZ02],H2AFY2,Q9P0M6,55506,22046132
-SARS ,SARS-CoV[GZ02],RPS20,P60866,6224,22046132
-SARS ,SARS-CoV[GZ02],CHEK2,O96017,11200,22046132
-SARS ,SARS-CoV[GZ02],TERF1,P54274,7013,22046132
-SARS ,SARS-CoV[GZ02],DCTN2,Q13561,10540,22046132
-SARS ,SARS-CoV[GZ02],DDX5,P17844,1655,22046132
-SARS ,SARS-CoV[GZ02],C11orf74,Q86VG3,119710,22046132
-SARS ,SARS-CoV[GZ02],EIF3F,O00303,8665,22046132
-SARS ,SARS-CoV[GZ02],EEF1A1,P68104,1915,22046132
-SARS ,SARS-CoV[GZ02],CAV1,Q03135,857,22046132
-SARS ,SARS-CoV[GZ02],IKBKB ,O14920,3551,22046132
-SARS ,SARS-CoV[GZ02],UBE2I,P63279,7329,22046132
-SARS ,SARS-CoV[GZ02],BTF3,P20290,689,22046132
-SARS ,SARS-CoV[GZ02],ATF5,Q9Y2D1,22809,22046132
-SARS ,SARS-CoV[GZ02],ND4L,P03901,4539,22046132
-SARS ,SARS-CoV[GZ02],COX2,P35354,4513,22046132
-SARS ,SARS-CoV[GZ02],HNRNPA1,P49312,3178,22046132
-SARS ,SARS-CoV[GZ02],ACE2,Q9BYF1,59272,22046132
-SARS ,SARS-CoV[GZ02],CLEC4G,Q6UXB4,339390,22046132
-SARS ,SARS-CoV[GZ02],CD209,Q9NNX6,30835,22046132
-SARS ,SARS-CoV[GZ02],CLEC4M,Q9H2X3,10332,22046132
-SARS ,SARS-CoV[GZ02],IRF3,Q14653,3661,22046132
-SARS ,SARS-CoV[GZ02],KPNA2,P52292,3838,22046132
-SARS ,SARS-CoV[GZ02],SFTPD ,P35247,6441,22046132
-SARS ,SARS-CoV[GZ02],PPIA ,P62937,5478,22046132
-SARS ,SARS-CoV[GZ02],BCL2L1,Q07817,598,22046132
-SARS ,SARS-CoV[GZ02],BCL2L2,Q92843,599,22046132
-SARS ,SARS-CoV[GZ02],MCL1,Q07820,4170,22046132
-SARS ,SARS-CoV[GZ02],BCL2A1,Q16548,597,22046132
-SARS ,SARS-CoV[GZ02],BCL2,P10415,596,22046132
-MERS ,SARS-CoV[GZ02],SKP2,Q13309,6502,31852899
-MERS ,SARS-CoV[GZ02],KPNA4,O00629,3840,29370303
-MERS ,SARS-CoV[GZ02],PRKRA ,O75569,8575,24522921
-MERS ,SARS-CoV[GZ02],CD9,P21926,928,28759649
-MERS ,SARS-CoV[GZ02],TMPRSS2,O15393,7113,28759649
-SARS ,SARS-CoV[GZ02],IKBKB ,O14920,3551,17705188
-HCoV-229E,HCoV-229E,ANPEP ,P15144,290,28643204
-HCoV-NL63,HCoV-NL63,ACE2,Q9BYF1,59272,28643204
-IBV ,IBV ,ZCRB1,Q8TBF4,85437,28643204
-MERS ,MERS-CoV[ChinaGD01],DPP4,P27487,1803,28643204
-MHV ,MHV ,HNRNPA1,P09651,3178,28643204
-MHV ,MHV ,SYNCRIP ,O60506,10492,28643204
-MHV ,MHV ,PTBP1,P26599,5725,28643204
-MHV ,MHV ,CEACAM1,P13688,634,28643204
-SARS ,SARS-CoV[GZ02],ZCRB1,Q8TBF4,85437,28643204
-SARS ,SARS-CoV[GZ02],ACE2,Q9BYF1,59272,28643204
-IBV ,IBV,ANXA2,P07355,302,28643204
-MHV ,MHV ,HNRNPA2B,P22626,3181,28643204
-MHV ,MHV ,HNRNPA3,P51991,220988,28643204
-MHV ,MHV ,ACO2,Q99798,50,28643204
-MHV ,MHV ,DNAJB1,P25685,3337,28643204
-MHV ,MHV ,HSPD1,P10809,3329,28643204
-MHV ,MHV ,HSPA9,P38646,3313,28643204
-MHV ,MHV ,COPB2,P35606,9276,30632963
-MHV ,MHV ,RPL13A,P40429,23521,30632963
-MHV ,MHV ,EIF3E,P60228,3646,30632963
-MHV ,MHV ,EIF3I,Q13347,8668,30632963
-MHV ,MHV ,NMT1,P30419,4836,30632963
-MHV ,MHV ,CHMP4B,Q9H444,128866,30632963
-MHV ,MHV ,EIF3F,O00303,8665,30632963
-MHV ,MHV ,GBF1,Q92538,8729,30632963
-MHV ,MHV ,RRM2,P31350,6241,30632963
-MHV ,MHV ,KIF11,P52732,3832,30632963
-MHV ,MHV ,PSMD1,Q99460,5707,30632963
-MHV ,MHV ,SRP54,P61011,6729,30632963
-MHV ,MHV ,NUDCD1,Q96RS6,84955,30632963
-MHV ,MHV ,NACA ,Q13765,4666,30632963
-MHV ,MHV ,SNX9,Q9Y5X1,51429,30632963
-MHV ,MHV ,BTF3,P20290,689,30632963
-MHV ,MHV ,SCFD1,Q8WVM8,23256,30632963
-MHV ,MHV ,PSMC2,P35998,5701,30632963
-MHV ,MHV ,TFEB ,P19484,7942,30632963
-MHV ,MHV ,TWF2,Q6IBS0,11344,30632963
-MHV ,MHV ,YKT6,O15498,10652,30632963
-MHV ,MHV ,KPNB1,Q14974,3837,30632963
-MHV ,MHV ,STX5,Q13190,6811,30632963
-MHV ,MHV ,STAT5A,P42229,6776,30632963
-MHV ,MHV ,RSL24D1,Q9UHA3,51187,30632963
-MHV ,MHV ,ACBD5,Q5T8D3,91452,30632963
-SARS ,SARS-CoV[GZ02],PHB ,P35232,5245,STRING Viruses 
-SARS ,SARS-CoV[GZ02],JUN ,P05412,3725,STRING Viruses 
-SARS ,SARS-CoV[GZ02],STAT3,P40763,6774,STRING Viruses 
-SARS ,SARS-CoV[GZ02],PPP1CA,P62136,5499,STRING Viruses 
-SARS ,SARS-CoV[GZ02],SPECC1,Q5M775,92521,STRING Viruses 
-SARS ,SARS-CoV[GZ02],FGL2,Q14314,10875,STRING Viruses 
-SARS ,SARS-CoV[GZ02],XPO1,O14980,7514,STRING Viruses 
-HCoV-229E,HCoV-229E,TGFB1,P01137,7040,STRING Viruses 
-HCoV-229E,HCoV-229E,FGL2,Q14314,10875,STRING Viruses 
-MHV ,MHV,SMAD3,P84022,4088,STRING Viruses
