--- Dataset/Reactome/INIT.txt
@@ -1,1 +1,0 @@
-DATASET
