+++ loader/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+# pylint: disable=invalid-name
+"""Define the migrators for different data sources."""
